Three 2FA demo/requirements.txt
Flask
Flask-Mail
itsdangerous
Werkzeug
twilio
python-dotenv
pyotp
qrcode[pil]
Pillow